Moved touch-enroll specs to e2e, added unit tests
* Use mocks and stubs for unit tests. Added test for getUserDefault
* Moved the others over to functional because they make changes to the SYSTEM
